pyproject.toml
[project]
name = "ultra-trail-strategist"
version = "0.1.0"
description = "Agentic AI Application for ultra-distance race strategies."
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "scipy>=1.10.0",
    "gpxpy>=1.5.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "scikit-learn>=1.3.0",
    "mcp>=0.1.0",
    "garminconnect",
    "fastmcp",
    "streamlit",
    "plotly",
    "watchdog",
    "requests-cache",
    "folium",
    "streamlit-folium",
    "python-telegram-bot"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "types-requests",
    "types-python-dateutil",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src"

